refactor lie group solver
