+++ doc/src/modules/solvers/ode.rst
@@ -102,12 +102,12 @@ factorable
 
 2nd_linear_airy
 ^^^^^^^^^^^^^^^
-.. autoclass:: sympy.solvers.ode.single::SecondLinearAiry
+.. autoclass:: sympy.solvers.ode.single::LinearAiry2nd
    :members:
 
 2nd_linear_bessel
 ^^^^^^^^^^^^^^^^^
-.. autoclass:: sympy.solvers.ode.single::SecondLinearBessel
+.. autoclass:: sympy.solvers.ode.single::LinearBessel2nd
    :members:
 
 Bernoulli
@@ -187,11 +187,12 @@ separable_reduced
 
 lie_group
 ^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::ode_lie_group
+.. autoclass:: sympy.solvers.ode.ode::LieGroup
+   :members:
 
 2nd_hypergeometric
 ^^^^^^^^^^^^^^^^^^
-.. autoclass:: sympy.solvers.ode.single::SecondHypergeometric
+.. autoclass:: sympy.solvers.ode.single::Hypergeometric2nd
    :members:
 
 1st_power_series
@@ -214,39 +215,39 @@ implemented for the various heuristics.
 
 abaco1_simple
 ^^^^^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_abaco1_simple
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_abaco1_simple
 
 abaco1_product
 ^^^^^^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_abaco1_product
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_abaco1_product
 
 bivariate
 ^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_bivariate
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_bivariate
 
 chi
 ^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_chi
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_chi
 
 abaco2_similar
 ^^^^^^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_abaco2_similar
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_abaco2_similar
 
 function_sum
 ^^^^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_function_sum
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_function_sum
 
 abaco2_unique_unknown
 ^^^^^^^^^^^^^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_abaco2_unique_unknown
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_abaco2_unique_unknown
 
 abaco2_unique_general
 ^^^^^^^^^^^^^^^^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_abaco2_unique_general
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_abaco2_unique_general
 
 linear
 ^^^^^^
-.. autofunction:: sympy.solvers.ode.ode::lie_heuristic_linear
+.. autofunction:: sympy.solvers.ode.lie_group::lie_heuristic_linear
 
 System of ODEs
 --------------
