+++ sympy/solvers/ode/__init__.py
@@ -1,5 +1,7 @@
 from .ode import (allhints, checkinfsol, classify_ode,
-        constantsimp, dsolve, homogeneous_order, infinitesimals)
+        constantsimp, dsolve, homogeneous_order)
+
+from .lie_group import infinitesimals
 
 from .subscheck import checkodesol
